requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
google-generativeai==0.7.2
PyPDF2==3.0.1
python-docx==1.1.2
Werkzeug==3.0.3
python-dotenv==1.0.1
Pillow==10.4.0
